+++ test/blunder/absinthe/error_handler/bug_snag_test.exs
@@ -4,36 +4,38 @@ if Code.ensure_loaded?(Bugsnag) do
     doctest Blunder.Absinthe.ErrorHandler.BugSnag
 
     import Mock
+
     test_with_mock "reports to Bugsnag with an original error when original error exists",
-                   Bugsnag, [report: fn (_, _) -> :ok end] do
+                   Bugsnag, [sync_report: fn (_, _) -> :ok end] do
       blunder = %Blunder{severity: :error, code: :code, original_error: "unexpected"}
       Blunder.Absinthe.ErrorHandler.BugSnag.call(blunder)
-      assert called Bugsnag.report(
+      assert called Bugsnag.sync_report(
         blunder.original_error, context: :code, metadata: :_, severity: "error", stacktrace: blunder.stacktrace
       )
     end
 
     test_with_mock "reports to Bugsnag with an blunder error when original error does not exist",
-                   Bugsnag, [report: fn (_, _) -> :ok end] do
+                   Bugsnag, [sync_report: fn (_, _) -> :ok end] do
       blunder = %Blunder{severity: :error, code: :code}
       Blunder.Absinthe.ErrorHandler.BugSnag.call(blunder)
-      assert called Bugsnag.report(
+      assert called Bugsnag.sync_report(
         blunder, context: :code, metadata: :_, severity: "error", stacktrace: blunder.stacktrace
       )
     end
 
     test_with_mock "reports to Bugsnag with an original error when original error is a tuple",
-                   Bugsnag, [report: fn (_, _) -> :ok end] do
+                   Bugsnag, [sync_report: fn (_, _) -> :ok end] do
       blunder = %Blunder{severity: :error, code: :code, original_error: {:error, "unexpected"}}
       Blunder.Absinthe.ErrorHandler.BugSnag.call(blunder)
-      assert called Bugsnag.report(
+      assert called Bugsnag.sync_report(
         elem(blunder.original_error, 1), context: :code, metadata: :_, severity: "error", stacktrace: blunder.stacktrace
       )
+    end
 
     test_with_mock "reports to Bugsnag", Bugsnag, [sync_report: fn (_, _) -> :ok end] do
       blunder = %Blunder{severity: :error, code: :code}
       Blunder.Absinthe.ErrorHandler.BugSnag.call(blunder)
-      assert called Bugsnag.sync_report(blunder, context: :code, metadata: :_, severity: "error")
+      assert called Bugsnag.sync_report(blunder, context: :code, metadata: :_, severity: "error", stacktrace: blunder.stacktrace)
     end
 
     test_with_mock "when Bugsnag returns error", Bugsnag, [sync_report: fn (_, _) -> :error end] do
@@ -51,7 +53,7 @@ if Code.ensure_loaded?(Bugsnag) do
       test_with_mock "test report is sent to bugsnag", Bugsnag, [sync_report: fn (_, _) -> :ok end] do
         blunder = %Blunder{severity: :error, code: :code}
         Blunder.Absinthe.ErrorHandler.BugSnag.call(blunder)
-        assert called Bugsnag.sync_report(blunder, context: :code, metadata: :_, severity: "error")
+        assert called Bugsnag.sync_report(blunder, context: :code, metadata: :_, severity: "error", stacktrace: blunder.stacktrace)
       end
 
       test_with_mock "when threshold is debug", Bugsnag, [sync_report: fn (_, _) -> :ok end] do