+++ lib/blunder/absinthe/error_handler.ex
@@ -3,28 +3,28 @@ defmodule Blunder.Absinthe.ErrorHandler do
   Behaviour for a module that handles Blunder errors in the Blunder.Absinthe middleware
 
   Creating an error handler is as simple as this:
-  
+
   ```elixir
   defmodule LogError do
     use Blunder.Absinthe.ErrorHandler
     require Logger
-  
+
     @impl Blunder.Absinthe.ErrorHandler
     def call(blunder) do
       Logger.error blunder.message
     end
-  
+
   end
   ```
-  
+
   Then in your config...
-  
+
   ```elixir
   config :blunder, error_handlers: [ LogError ]
   ```
   """
 
-  @callback call(Blunder.t) :: (:ok | {:error, any})
+  @callback call(Blunder.t, any) :: (:ok | {:error, any})
 
   defmacro __using__(_opts \\ []) do
     quote do
@@ -37,13 +37,13 @@ defmodule Blunder.Absinthe.ErrorHandler do
         {:ok, args}
       end
 
-      def start_link(_ \\ nil) do
-        GenServer.start_link(__MODULE__, nil, name: __MODULE__)
+      def start_link(args) do
+        GenServer.start_link(__MODULE__, args, name: __MODULE__)
       end
 
       @impl GenServer
       def handle_cast({:error_event, blunder}, state) do
-        call(blunder)
+        call(blunder, state)
         {:noreply, state}
       end
     end